[2803] Fix subjects namepace
