+++ db/migrate/20191024143538_add_financial_incentive_to_subjects.rb
@@ -1,7 +1,7 @@
 class AddFinancialIncentiveToSubjects < ActiveRecord::Migration[6.0]
   def up
     say_with_time "populating subjects finanical" do
-      SubjectFinancialIncentiveCreatorService.new.execute
+      Subjects::FinancialIncentiveCreatorService.new.execute
     end
   end
