+++ db/migrate/20200115141035_add_subject_knowledge_enhancement_available_to_financial_incentive.rb
@@ -3,7 +3,7 @@ class AddSubjectKnowledgeEnhancementAvailableToFinancialIncentive < ActiveRecord
     add_column :financial_incentive, :subject_knowledge_enhancement_course_available, :boolean, null: false, default: false
 
     say_with_time "populating finanical incentive subject knowledge enhancement course available" do
-      SubjectFinancialIncentiveSetSubjectKnowledgeEnhancementCourseAvailableService.new.execute
+      Subjects::FinancialIncentiveSetSubjectKnowledgeEnhancementCourseAvailableService.new.execute
     end
   end
 end
