+++ spec/support/reference_data.rb
@@ -23,10 +23,9 @@ RSpec.configure do |config|
   # the tag "without_subjects: true" to make sure the subjects table is empty
   # for tests that need it.
   config.before(:all) do
-    SubjectCreatorService.new.execute
-    SubjectFinancialIncentiveCreatorService.new.execute
-    SubjectFinancialIncentiveSetSubjectKnowledgeEnhancementCourseAvailableService.new.execute
-
+    Subjects::CreatorService.new.execute
+    Subjects::FinancialIncentiveCreatorService.new.execute
+    Subjects::FinancialIncentiveSetSubjectKnowledgeEnhancementCourseAvailableService.new.execute
     # Creating or finding course test data
     if Settings.test_setup
       TestSetup.new
